requirements.txt
langchain_google_genai
transformers
langchain
langchain-core 
langchain_google_genai
sentence-transformers==2.2.2
InstructorEmbedding
faiss-cpu
langchain_community
streamlit
python-dotenv
schedule